WIP: Alter scan specs so that all have balanced list specifiers
The original pn_data_t scan code didn't require balanced lists as it
worked char by char. The new generator requires balanced list
specifiers. This is back compatible as the old scan code doesn't try to
reconcile the number of items to be read with the actual number of items
in the list.