WIP: Initial cut at a cffi library for proton
Build an extension library cproton_ffi which can be used by a cproton
module to interface to libqpid_proton_core.

The intention is that the higher level code continure ot use cproton
exactly as before but the cproton module is now written manually to
interface to the ffi code instead if being written by swig.