PROTON-2095: Use CFFI instead of SWIG for python binding
Build an extension library cproton_ffi which can be used by a cproton
module to interface to libqpid_proton_core.

The intention is that the higher level code continue to use cproton
as before but the cproton module is now written manually to
interface to the ffi code instead if being written by swig.