PROTON-????: Use the new transport APIs in the connection driver
This needed some changes in the test code that uses the connection
driver. This is because this code assumes that some things either happen
in a specific order due to the previous implementation, or that the new
connection driver using code needs to keep on asking the connection
driver for more bytes to write until there are none to satisfy the test
assumptions (this isn't a bad thing as it is as a general pattern more
efficient anyway.